+++ src/react/ItemsHtmlLayer.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import { Items, Item } from "../schema/app_schema.js";
+import { ItemView } from "./itemux.js";
+
+export function ItemsHtmlLayer(props: {
+	items: Items;
+	canvasPosition: { left: number; top: number };
+	pan: { x: number; y: number };
+	zoom: number;
+}): JSX.Element {
+	const { items, canvasPosition, pan, zoom } = props;
+	return (
+		<div className="relative h-full w-full" style={{ left: `${pan.x}px`, top: `${pan.y}px`, transform: `scale(${zoom})`, transformOrigin: "0 0" }}>
+			{items.map((item, index) =>
+				item instanceof Item ? (
+					<ItemView
+						item={item}
+						key={item.id}
+						index={index}
+						canvasPosition={{ left: canvasPosition.left, top: canvasPosition.top }}
+						hideSelectionControls
+						pan={pan}
+						zoom={zoom}
+					/>
+				) : (
+					<></>
+				)
+			)}
+		</div>
+	);
+}
